refactor: restructures landing page folders to be consistent
