+++ application/next-frontend/src/Landing/scenarios/FifthSection/components/ImageWithPostit.tsx
