+++ application/next-frontend/src/Landing/scenarios/FirstSection/components/madeByBlank.tsx
