+++ application/next-frontend/src/Landing/scenarios/FirstSection/components/quotesBubbles.tsx
