+++ application/next-frontend/src/Landing/scenarios/SecondSection/components/ImageContainer.tsx
